feat(platform): remove name/tab watchers. fixup browser prop fetch
