+++ src/platform/src/events/mod.rs
@@ -1,11 +1,7 @@
 mod foreground;
 mod interaction;
-mod name;
 mod system;
-mod tab;
 
 pub use foreground::*;
 pub use interaction::*;
-pub use name::*;
 pub use system::*;
-pub use tab::*;
